Custom styling
- Created new `@uploadthing/tw` package to contain utility function for TailwindCSS config adjustment. This package was created in such way that it can be easily imported into `tailwind.config` file
- Adjusted example
- Removed `tw.ts` from react package
