+++ examples/appdir/package.json
@@ -32,6 +32,7 @@
     "@uploadthing/eslint-config": "0.1.0",
     "@uploadthing/tsconfig": "0.1.0",
     "eslint": "^8.42.0",
-    "typescript": "5.1.3"
+    "typescript": "5.1.3",
+    "@uploadthing/tw": "5.1.0"
   }
 }
