+++ packages/react/tsup.config.ts
@@ -4,7 +4,7 @@ import { config } from "@uploadthing/tsup-config";
 
 export default defineConfig((opts) => ({
   ...config,
-  entry: ["./src/index.ts", "./src/hooks.ts", "./src/tw.ts"],
+  entry: ["./src/index.ts", "./src/hooks.ts"],
   clean: !opts.watch,
   esbuildOptions: (option) => {
     option.banner = {
