+++ packages/tw/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@uploadthing/tw",
+  "version": "5.1.0",
+  "description": "",
+  "main": "index.js",
+  "types": "types/index.d.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "tailwindcss": "^3.3.2",
+    "typescript": "5.1.3"
+  }
+}
