+++ packages/tw/types/index.d.ts
@@ -1,0 +1,5 @@
+import type { Config } from 'tailwindcss'
+
+declare const withUt: (twConfig: Config) => Config
+
+export default withUt
